backend-ssE/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.5.0
python-dotenv==1.0.0
Werkzeug==2.3.7
bson==0.5.10
python-multipart==0.0.6
Pillow==10.0.1
geopy==2.4.0